backend/auth-service/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic==2.4.2
pydantic-settings==2.0.3

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.28.0

# HTTP Client (for OAuth2)
httpx==0.25.1

# Redis (for token blacklisting and rate limiting)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
asgi-lifespan==2.1.0 